x trying to find and correct bugs detected by the compiler and/or Coverty
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@6322 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/File__Analyze_Buffer.cpp
@@ -1682,7 +1682,7 @@ void File__Analyze::Get_ISO_6937_2(int64u Bytes, Ztring &Info, const char* Name)
 {
     INTEGRITY_SIZE_ATLEAST_STRING(Bytes);
     Info.clear();
-    size_t End=Buffer_Offset+(size_t)Element_Offset+Bytes;
+    size_t End = Buffer_Offset + (size_t)Element_Offset + (size_t)Bytes;
     for (size_t Pos=Buffer_Offset+(size_t)Element_Offset; Pos<End; ++Pos)
     {
         wchar_t EscapeChar=__T('\x0000');
@@ -1807,7 +1807,7 @@ void File__Analyze::Get_ISO_8859_5(int64u Bytes, Ztring &Info, const char* Name)
 {
     INTEGRITY_SIZE_ATLEAST_STRING(Bytes);
     Info.clear();
-    size_t End=Buffer_Offset+(size_t)Element_Offset+Bytes;
+    size_t End = Buffer_Offset + (size_t)Element_Offset + (size_t)Bytes;
     for (size_t Pos=Buffer_Offset+(size_t)Element_Offset; Pos<End; ++Pos)
     {
         switch (Buffer[Pos])
